OvaCare-main/pcos-ml-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
tensorflow>=2.16.0
keras>=3.0.0
